+++ tools/generator/ui/src/main/java/is/codion/tools/generator/ui/DomainGeneratorPanel.java
@@ -177,7 +177,7 @@ public final class DomainGeneratorPanel extends JPanel {
 										.control(Controls.builder()
 														.name("Columns")
 														.control(table.createToggleColumnsControls())
-														.control(table.createAutoResizeModeControl()))
+														.control(table.createSelectAutoResizeModeControl()))
 										.build())
 						.onBuild(table -> table.sortModel().setSortOrder(SchemaColumns.Id.SCHEMA, SortOrder.ASCENDING))
 						.build();
@@ -186,7 +186,7 @@ public final class DomainGeneratorPanel extends JPanel {
 	private FilterTable<EntityRow, EntityColumns.Id> createEntityTable() {
 		return FilterTable.builder(model.entityModel(), createEntityColumns())
 						.autoResizeMode(JTable.AUTO_RESIZE_ALL_COLUMNS)
-						.popupMenuControl(FilterTable::createAutoResizeModeControl)
+						.popupMenuControl(FilterTable::createSelectAutoResizeModeControl)
 						.onBuild(table -> table.sortModel().setSortOrder(EntityColumns.Id.ENTITY, SortOrder.ASCENDING))
 						.build();
 	}