+++ tools/loadtest/ui/src/main/java/is/codion/tools/loadtest/ui/LoadTestPanel.java
@@ -402,7 +402,7 @@ public final class LoadTestPanel<T> extends JPanel {
 														.name("Columns")
 														.control(table.createToggleColumnsControls())
 														.control(table.createResetColumnsControl())
-														.control(table.createAutoResizeModeControl()))
+														.control(table.createSelectAutoResizeModeControl()))
 										.build())
 						.build();
 	}