+++ tools/monitor/ui/src/main/java/is/codion/tools/monitor/ui/ClientMonitorPanel.java
@@ -144,7 +144,7 @@ public final class ClientMonitorPanel extends JPanel {
 										.name("Columns")
 										.control(table.createToggleColumnsControls())
 										.control(table.createResetColumnsControl())
-										.control(table.createAutoResizeModeControl())))
+										.control(table.createSelectAutoResizeModeControl())))
 						.createPopupMenu();
 	}